First attempt to switch to packagable version
