+++ .flake8
@@ -1,2 +1,3 @@
 [flake8]
 max-line-length = 160
+per-file-ignores = __init__.py:F401
