+++ .pre-commit-config.yaml
@@ -1,4 +1,11 @@
 repos:
+-   repo: https://github.com/pre-commit/pre-commit-hooks
+    rev: v4.3.0
+    hooks:
+    -   id: check-yaml
+    -   id: end-of-file-fixer
+    -   id: trailing-whitespace
+    -   id: mixed-line-ending
 -   repo: https://github.com/psf/black
     rev: 22.6.0
     hooks:
