+++ ldapsearch-ad.py
@@ -1,858 +1,19 @@
 #!/usr/bin/env python3
 
 import argparse
-import datetime
-import ldap3
 import logging
-import re
-import sys
-import struct
-import codecs
 
+from sys import exit
+from sys import stdout
 
-def print_version():
-    # version number is just the date of the release.
-    # with format: YYYY.MM.DD
-    version_number = "2022.08.17"
-    print(f"ldapsearch-ad v{version_number}")
+from ldapsearchad import LdapsearchAd
 
+from ldapsearchad.logging import log_title
+from ldapsearchad.logging import log_error
+from ldapsearchad.logging import log_info
 
-def c_red(message):
-    """Color text for bad configuration."""
-    return f"\x1b[0;31;40m{message}\x1b[0m"
-
-
-def c_green(message):
-    """Color text for good configuration."""
-    return f"\x1b[0;32;40m{message}\x1b[0m"
-
-
-def c_orange(message):
-    """Color text for weak configuration."""
-    return f"\x1b[0;33;40m{message}\x1b[0m"
-
-
-def c_blue(message):
-    """Color text for information."""
-    return f"\x1b[0;34;40m{message}\x1b[0m"
-
-
-def c_purple(message):
-    """Color text for abnormal behavior of the tool itself."""
-    return f"\x1b[0;35;40m{message}\x1b[0m"
-
-
-def c_cyan(message):
-    """Mostly for general usefull information."""
-    return f"\x1b[0;36;40m{message}\x1b[0m"
-
-
-def c_white_on_red(message):
-    """Color text for very bad configuration."""
-    return f"\x1b[1;37;41m{message}\x1b[0m"
-
-
-def log_title(title, level=2):
-    heading = "#" * level
-    logging.info(f"\x1b[1;37;40m{heading} {title} {heading}\x1b[0m")
-
-
-def log_error(message):
-    logging.error(f'{c_red("[-]")} {message}')
-
-
-def log_warning(message):
-    logging.info(f'{c_orange("[!]")} {message}')
-
-
-def log_info(message):
-    logging.info(f'{c_blue("[+]")} {message}')
-
-
-def log_success(message):
-    logging.info(f'{c_green("[*]")} {message}')
-
-
-def str_human_date(date):
-    # ldap3 version 2.6 returns a datetime.timedelta object
-    if isinstance(date, datetime.timedelta):
-        nb_sec = int(date.total_seconds())
-    # older versions return a negative big number
-    else:
-        nb_sec = int((-date) / 10000000)
-    if nb_sec > 60:
-        nb_min = int(nb_sec / 60)
-        nb_sec = nb_sec % 60
-        if nb_min > 60:
-            nb_hour = int(nb_min / 60)
-            nb_min = nb_min % 60
-            if nb_hour > 24:
-                nb_day = int(nb_hour / 24)
-                nb_hour = nb_hour % 24
-                return f"{nb_day} days, {nb_hour} hours, {nb_min} minutes, {nb_sec} seconds"
-            return f"{nb_hour} hours, {nb_min} minutes, {nb_sec} seconds"
-        return f"{nb_min} minutes, {nb_sec} seconds"
-    return f"{nb_sec} seconds"
-
-
-def str_functionality_level(num):
-    """Return the functionality level as described at:
-    https://msdn.microsoft.com/en-us/library/cc223274.aspx
-    Note: it is the same for forest, domain, and domain controller."""
-    n = int(num)
-    func_levels = [
-        c_white_on_red("Windows 2000"),
-        c_white_on_red("Windows 2003 with mixed domains"),
-        c_white_on_red("Windows 2003"),
-        c_red("Windows 2008"),
-        c_red("Windows 2008 R2"),
-        c_orange("Windows 2012"),
-        c_orange("Windows 2012 R2"),
-        c_green("Windows 2016"),
-    ]
-    if 0 <= n < len(func_levels):
-        return func_levels[n]
-    else:
-        return f"Not known, update this script. (value = {num})"
-
-
-def list_uac_flags(uac):
-    """Return a list of property flags as described at:
-    https://support.microsoft.com/en-gb/help/305144/how-to-use-the-useraccountcontrol-flags-to-manipulate-user-account-pro"""
-    flags = []
-    if uac & 0x1 > 0:
-        flags.append("SCRIPT")
-    if uac & 0x2 > 0:
-        flags.append("ACCOUNTDISABLE")
-    if uac & 0x8 > 0:
-        flags.append("HOMEDIR_REQUIRED")
-    if uac & 0x10 > 0:
-        flags.append("LOCKOUT")
-    if uac & 0x20 > 0:
-        flags.append("PASSWD_NOTREQD")
-    if uac & 0x40 > 0:
-        flags.append("PASSWD_CANT_CHANGE")
-    if uac & 0x80 > 0:
-        flags.append("ENCRYPTED_TEXT_PWD_ALLOWED")
-    if uac & 0x100 > 0:
-        flags.append("TEMP_DUPLICATE_ACCOUNT")
-    if uac & 0x200 > 0:
-        flags.append("NORMAL_ACCOUNT")
-    if uac & 0x800 > 0:
-        flags.append("INTERDOMAIN_TRUST_ACCOUNT")
-    if uac & 0x1000 > 0:
-        flags.append("WORKSTATION_TRUST_ACCOUNT")
-    if uac & 0x2000 > 0:
-        flags.append("SERVER_TRUST_ACCOUNT")
-    if uac & 0x10000 > 0:
-        flags.append("DONT_EXPIRE_PASSWORD")
-    if uac & 0x20000 > 0:
-        flags.append("MNS_LOGON_ACCOUNT")
-    if uac & 0x40000 > 0:
-        flags.append("SMARTCARD_REQUIRED")
-    if uac & 0x80000 > 0:
-        flags.append("TRUSTED_FOR_DELEGATION")
-    if uac & 0x100000 > 0:
-        flags.append("NOT_DELEGATED")
-    if uac & 0x200000 > 0:
-        flags.append("USE_DES_KEY_ONLY")
-    if uac & 0x400000 > 0:
-        flags.append("DONT_REQ_PREAUTH")
-    if uac & 0x800000 > 0:
-        flags.append("PASSWORD_EXPIRED")
-    if uac & 0x1000000 > 0:
-        flags.append("TRUSTED_TO_AUTH_FOR_DELEGATION")
-    if uac & 0x04000000 > 0:
-        flags.append("PARTIAL_SECRETS_ACCOUNT")
-    return flags
-
-
-def list_uac_colored_flags(uac):
-    """Return a list of property flags as described at:
-    https://support.microsoft.com/en-gb/help/305144/how-to-use-the-useraccountcontrol-flags-to-manipulate-user-account-pro"""
-    flags = []
-    if uac & 0x1 > 0:
-        flags.append("SCRIPT")
-    if uac & 0x2 > 0:
-        flags.append(c_cyan("ACCOUNTDISABLE"))
-    if uac & 0x8 > 0:
-        flags.append("HOMEDIR_REQUIRED")
-    if uac & 0x10 > 0:
-        flags.append(c_orange("LOCKOUT"))
-    if uac & 0x20 > 0:
-        flags.append(c_white_on_red("PASSWD_NOTREQD"))
-    if uac & 0x40 > 0:
-        flags.append(c_red("PASSWD_CANT_CHANGE"))
-    if uac & 0x80 > 0:
-        flags.append(c_white_on_red("ENCRYPTED_TEXT_PWD_ALLOWED"))
-    if uac & 0x100 > 0:
-        flags.append("TEMP_DUPLICATE_ACCOUNT")
-    if uac & 0x200 > 0:
-        flags.append("NORMAL_ACCOUNT")
-    if uac & 0x800 > 0:
-        flags.append(c_cyan("INTERDOMAIN_TRUST_ACCOUNT"))
-    if uac & 0x1000 > 0:
-        flags.append(c_cyan("WORKSTATION_TRUST_ACCOUNT"))
-    if uac & 0x2000 > 0:
-        flags.append(c_cyan("SERVER_TRUST_ACCOUNT"))
-    if uac & 0x10000 > 0:
-        flags.append(c_red("DONT_EXPIRE_PASSWORD"))
-    if uac & 0x20000 > 0:
-        flags.append("MNS_LOGON_ACCOUNT")
-    if uac & 0x40000 > 0:
-        flags.append("SMARTCARD_REQUIRED")
-    if uac & 0x80000 > 0:
-        flags.append(c_orange("TRUSTED_FOR_DELEGATION"))
-    if uac & 0x100000 > 0:
-        flags.append("NOT_DELEGATED")
-    if uac & 0x200000 > 0:
-        flags.append(c_red("USE_DES_KEY_ONLY"))
-    if uac & 0x400000 > 0:
-        flags.append(c_red("DONT_REQ_PREAUTH"))
-    if uac & 0x800000 > 0:
-        flags.append(c_cyan("PASSWORD_EXPIRED"))
-    if uac & 0x1000000 > 0:
-        flags.append(c_orange("TRUSTED_TO_AUTH_FOR_DELEGATION"))
-    if uac & 0x04000000 > 0:
-        flags.append("PARTIAL_SECRETS_ACCOUNT")
-    return flags
-
-
-def str_samaccounttype(sat):
-    """Return the SAM-Account-Type as described at:
-    https://docs.microsoft.com/en-us/windows/desktop/adschema/a-samaccounttype"""
-    if sat == 0x0:
-        return "SAM_DOMAIN_OBJECT"
-    elif sat == 0x10000000:
-        return "SAM_GROUP_OBJECT"
-    elif sat == 0x10000001:
-        return "SAM_NON_SECURITY_GROUP_OBJECT"
-    elif sat == 0x20000000:
-        return "SAM_ALIAS_OBJECT"
-    elif sat == 0x20000001:
-        return "SAM_NON_SECURITY_ALIAS_OBJECT"
-    elif sat == 0x30000000:
-        return "SAM_USER_OBJECT"
-    elif sat == 0x30000000:
-        return "SAM_NORMAL_USER_ACCOUNT"
-    elif sat == 0x30000001:
-        return "SAM_MACHINE_ACCOUNT"
-    elif sat == 0x30000002:
-        return "SAM_TRUST_ACCOUNT"
-    elif sat == 0x40000000:
-        return "SAM_APP_BASIC_GROUP"
-    elif sat == 0x40000001:
-        return "SAM_APP_QUERY_GROUP"
-    elif sat == 0x7FFFFFFF:
-        return "SAM_ACCOUNT_TYPE_MAX"
-    else:
-        return c_purple("Error: unknown value")
-
-
-def str_object_type(entry):
-    if "sAMAccountType" in entry:
-        sat = entry["sAMAccountType"]
-        if sat == 0x0:
-            return "domain"
-        elif sat == 0x10000000:
-            return "group"
-        elif sat == 0x30000000:
-            return "user"
-        elif sat == 0x30000001:
-            return "computer"
-        else:
-            return c_purple(f"sAMAccountType = {sat}. Please complete this script.")
-    else:
-        return c_purple("Unable to find correct type (sAMAccountType not present).")
-
-
-def list_trust_type(trust_type):
-    """Return the trust type as defined here: https://msdn.microsoft.com/en-us/library/cc223771.aspx"""
-    if trust_type == 1:
-        return "The trusted domain is a Windows domain not running Active Directory."
-    elif trust_type == 2:
-        return "The trusted domain is a Windows domain running Active Directory."
-    elif trust_type == 3:
-        return "The trusted domain is running a non-Windows, RFC4120-compliant Kerberos distribution."
-    elif trust_type == 4:
-        return "Historical reference; this value is not used in Windows."
-    else:
-        return "Error: unknown value."
-
-
-def list_trust_direction(trust_direction):
-    """Return the trust direction as defined here: https://msdn.microsoft.com/en-us/library/cc223768.aspx"""
-    if trust_direction == 0:
-        return "Disabled"
-    elif trust_direction == 1:
-        return c_green("Outbound")
-    elif trust_direction == 2:
-        return c_green("Inbound")
-    elif trust_direction == 3:
-        return c_cyan("Bidirectional")
-    else:
-        return c_purple("Error: unknown value.")
-
-
-def list_trust_attributes(ta):
-    """Return the trust attribute flags as defined here: https://msdn.microsoft.com/en-us/library/cc223779.aspx"""
-    flags = []
-    if ta & 0x1 > 0:
-        flags.append("TRUST_ATTRIBUTE_NON_TRANSITIVE")
-    if ta & 0x2 > 0:
-        flags.append("TRUST_ATTRIBUTE_UPLEVEL_ONLY")
-    if ta & 0x4 > 0:
-        flags.append("TRUST_ATTRIBUTE_QUARANTINED_DOMAIN")
-    if ta & 0x8 > 0:
-        flags.append("TRUST_ATTRIBUTE_FOREST_TRANSITIVE")
-    if ta & 0x10 > 0:
-        flags.append("TRUST_ATTRIBUTE_CROSS_ORGANIZATION")
-    if ta & 0x20 > 0:
-        flags.append("TRUST_ATTRIBUTE_WITHIN_FOREST")
-    if ta & 0x40 > 0:
-        flags.append("TRUST_ATTRIBUTE_TREAT_AS_EXTERNAL")
-    if ta & 0x80 > 0:
-        flags.append("TRUST_ATTRIBUTE_USES_RC4_ENCRYPTION")
-    if ta & 0x200 > 0:
-        flags.append("TRUST_ATTRIBUTE_CROSS_ORGANIZATION_NO_TGT_DELEGATION")
-    if ta & 0x400 > 0:
-        flags.append("TRUST_ATTRIBUTE_PIM_TRUST")
-    return flags
-
-
-def debug_var(var):
-    print("\n===================")
-    print("===   Debug var  ===")
-    print("====================")
-    print(f"type of var = {type(var)}")
-    print("====================")
-    if isinstance(var, dict):
-        print("var is a dict:")
-        for k in var:
-            print(f"* {k} = {var[k]}")
-        print("====================")
-    if isinstance(var, list):
-        print("var is a list:")
-        for i in var:
-            print(f"* {i}")
-        print("====================")
-    if isinstance(var, str):
-        print("var is a str:")
-        print(var)
-        print("====================")
-    for i in sorted(var.__dir__()):
-        print(i)
-    print("====================")
-
-
-def convert_sid_to_string(hexstr):
-    value = codecs.encode(bytes(hexstr[0]), "hex")
-    init_value = bytes(hexstr[0])
-    value = "S-1-5"
-    for i in range(8, len(init_value), 4):
-        value += f'-{str(struct.unpack("<I", init_value[i:i+4])[0])}'
-    return value
-
-
-class LdapsearchAd:
-
-    hostname = None
-    domain = None
-    username = None
-    password = None
-    server = None
-    connection = None
-
-    last_errors = []
-
-    def __init__(
-        self,
-        hostname,
-        ssl=False,
-        domain=None,
-        username=None,
-        password=None,
-        hashes=None,
-    ):
-        self.hostname = hostname
-        self.domain = domain
-        self.username = username
-        self.password = password
-        self.hashes = hashes
-        if self.hashes is not None:
-            self.lmhash, self.nthash = self.hashes.split(":")
-
-        try:
-            self.server = ldap3.Server(self.hostname, use_ssl=ssl, get_info="ALL")
-            if self.domain and self.username and self.password:
-                self.connection = ldap3.Connection(
-                    self.server,
-                    user=f"{self.domain}\\{self.username}",
-                    password=self.password,
-                    authentication="NTLM",
-                    auto_bind=True,
-                )
-            elif self.hashes is not None:
-                if self.lmhash == "":
-                    self.lmhash = "aad3b435b51404eeaad3b435b51404ee"
-                self.connection = ldap3.Connection(
-                    self.server,
-                    user=f"{self.domain}\\{self.username}",
-                    password=self.lmhash + ":" + self.nthash,
-                    authentication=ldap3.NTLM,
-                    auto_bind=True,
-                )
-            else:
-                self.connection = ldap3.Connection(self.server, auto_bind=True)
-        except ldap3.core.exceptions.LDAPSocketOpenError:
-            self.last_errors.append(f"Unable to connect to {self.hostname}.")
-        except ldap3.core.exceptions.LDAPBindError as ldapbe:
-            self.last_errors.append(f"Bind Error: {ldapbe}")
-
-    def print_info(self):
-        log_info(
-            f'Forest functionality level = {str_functionality_level(self.server.info.other["forestFunctionality"][0])}'
-        )
-        log_info(
-            f'Domain functionality level = {str_functionality_level(self.server.info.other["domainFunctionality"][0])}'
-        )
-        log_info(
-            f'Domain controller functionality level = {str_functionality_level(self.server.info.other["domainControllerFunctionality"][0])}'
-        )
-        log_info(
-            f'rootDomainNamingContext = {self.server.info.other["rootDomainNamingContext"][0]}'
-        )
-        log_info(
-            f'defaultNamingContext = {self.server.info.other["defaultNamingContext"][0]}'
-        )
-        log_info(f'ldapServiceName = {self.server.info.other["ldapServiceName"][0]}')
-        log_info(f"naming_contexts = {self.server.info.naming_contexts}")
-
-    def whoami(self):
-        if self.server and self.connection:
-            log_info(self.connection.extend.standard.who_am_i())
-        else:
-            log_error(
-                "Not connected to LDAP server. Please authenticate yourself before trying to search something."
-            )
-
-    def old_search(self, search_filter, attributes="*", size_limit=100):
-        """This function remains for backward compatibility only.
-        It is usually limited to the first page of result (which returns 1000 results)
-        against an Active Directory domain controller.
-        You should use search() instead."""
-        r = []
-        if self.server and self.connection:
-            base_dn = self.server.info.other.get("defaultNamingContext")[0]
-            if search_filter[0] != "(":
-                search_filter = f"({search_filter})"
-            try:
-                self.connection.search(
-                    base_dn, search_filter, attributes=attributes, size_limit=size_limit
-                )
-                entries = self.connection.entries
-                for entry in entries:
-                    r.append(entry)
-            except ldap3.core.exceptions.LDAPInvalidFilterError as e:
-                log_error(f"{e} (perhaps missing parenthesis?)")
-        else:
-            log_error(
-                "Not connected to LDAP server. Please authenticate yourself before trying to search something."
-            )
-            sys.exit(4)
-        return r
-
-    def search(self, search_filter, attributes="*", size_limit=100, page_size=1000):
-        """Warning: returns a generator!
-        Thus, it is not backward compatible with the old_search
-        Arguments:
-        * search_filter = '(samaccountname=adm*)'
-        * attributes = ['cn', 'givenName'] or '*' for all attributes
-        Return value:
-        * a generator of CaseInsensitiveDict"""
-
-        nb_entries = 0
-        if self.server and self.connection:
-            base_dn = self.server.info.other.get("defaultNamingContext")[0]
-            self.connection.search(
-                search_base=base_dn,
-                search_filter=search_filter,
-                search_scope=ldap3.SUBTREE,
-                attributes=attributes,
-                paged_size=page_size,
-                size_limit=size_limit,
-            )
-            for entry in self.connection.response:
-                if entry["type"] == "searchResEntry":
-                    nb_entries += 1
-                    yield entry["attributes"]
-            if "controls" in self.connection.result:
-                cookie = self.connection.result["controls"]["1.2.840.113556.1.4.319"][
-                    "value"
-                ]["cookie"]
-                size_limit -= nb_entries
-                nb_entries = 0
-            else:
-                size_limit = 0
-                cookie = None
-            while size_limit > 0 and cookie:
-                self.connection.search(
-                    search_base=base_dn,
-                    search_filter=search_filter,
-                    search_scope=ldap3.SUBTREE,
-                    attributes=attributes,
-                    paged_size=page_size,
-                    size_limit=size_limit,
-                    paged_cookie=cookie,
-                )
-                for entry in self.connection.response:
-                    if entry["type"] == "searchResEntry":
-                        nb_entries += 1
-                        yield entry["attributes"]
-                if "controls" in self.connection.result:
-                    cookie = self.connection.result["controls"][
-                        "1.2.840.113556.1.4.319"
-                    ]["value"]["cookie"]
-                    size_limit -= nb_entries
-                    nb_entries = 0
-                else:
-                    size_limit = 0
-                    cookie = None
-        else:
-            log_error(
-                "Not connected to LDAP server. Please authenticate yourself before trying to search something."
-            )
-            sys.exit(4)
-
-    def __list_groups(self, entry):
-        """Return a list containing the CN of each group the parameter is member of."""
-        if "memberOf" not in entry:
-            return ["memberOf attribute not found"]
-        groups = []
-        # dirty patch because "memberOf.value" return a string if there is only one group
-        # and a list of string if there is more than one group
-        logging.debug(f'type of memberOf = {type(entry["memberOf"])}')
-        if isinstance(entry["memberOf"], str):
-            groups_raw = [entry["memberOf"]]
-        else:
-            groups_raw = entry["memberOf"]
-        for group in groups_raw:
-            group_cn = re.search("CN=([^,]*),", group).group(1)
-            if re.search("^(enterprise admins)$", group_cn, re.IGNORECASE):
-                groups.append(c_red(group_cn))
-            elif re.search(
-                "^(domain admins|admins du domaine)$", group_cn, re.IGNORECASE
-            ):
-                groups.append(c_red(group_cn))
-            elif re.search(
-                "^(administrators|administrateurs)$", group_cn, re.IGNORECASE
-            ):
-                groups.append(c_red(group_cn))
-            elif re.search("admin", group_cn, re.IGNORECASE):
-                groups.append(c_orange(group_cn))
-            else:
-                groups.append(group_cn)
-        return groups
-
-    def __print_user_details(self, user, tab=""):
-        """Print info of a user (samacountname and userAccountControl)."""
-        log_info(f'{tab}{user["samAccountName"]}')
-        log_info(f"{tab}|___type: {str_object_type(user)}")
-        if "displayName" in user:
-            log_info(f'{tab}|___displayName = {user["displayName"]}')
-        if "description" in user:
-            log_info(f'{tab}|___description = {user["description"]}')
-        if "adminCount" in user:
-            if user["admincount"] == 1:
-                log_success(f'{tab}|___{c_red("The adminCount is set to 1")}')
-            elif user["admincount"] == 0:
-                log_info(f'{tab}|___{"adminCount = 0"}')
-            else:
-                log_error(
-                    f'{tab}|___Unknown value for adminCount: {user["admincount"]}'
-                )
-        if "userAccountControl" in user:
-            log_info(
-                f'{tab}|___userAccountControl = {", ".join(list_uac_colored_flags(user["userAccountControl"]))}'
-            )
-        log_info(
-            f'{tab}|___sAMAccountType = {str_samaccounttype(user["samaccounttype"])}'
-        )
-        if "memberOf" in user:
-            log_info(f'{tab}|___memberOf = {", ".join(self.__list_groups(user))}')
-
-    def __print_user_brief(self, user, tab=""):
-        """Print info of a user on a single line (samacountname and userAccountControl)."""
-        if str_object_type(user) != "user":
-            if "foreignSecurityPrincipal" in user["objectclass"]:
-                log_info(
-                    f'{tab}{user["name"]} (objectclass = foreignSecurityPrincipal, probably a user from another domain. Please investigate)'
-                )
-            else:
-                log_warning(f'{tab}{user["name"]} ({c_orange(str_object_type(user))})')
-        else:
-            uac_flags = list_uac_colored_flags(user["userAccountControl"])
-            uac_flags.remove("NORMAL_ACCOUNT")
-            if uac_flags:
-                log_success(f'{tab}{user["sAMAccountName"]} ({", ".join(uac_flags)})')
-            else:
-                log_info(f'{tab}{user["sAMAccountName"]}')
-
-    def print_users(self, search_filter, attributes="*", size_limit=100):
-        """Method to pretty print a set a users attributes."""
-        r_search = self.search(search_filter, attributes, size_limit=size_limit)
-        for user in r_search:
-            self.__print_user_details(user)
-
-    def print_users_list(self, search_filter, attributes="*", size_limit=100):
-        """Method to pretty list print a set a users attributes."""
-        r_search = self.search(search_filter, attributes, size_limit=size_limit)
-        for user in r_search:
-            self.__print_user_brief(user)
-
-    def print_trusts(self):
-        """Method to get infos about trusts."""
-        for trust in self.search("(objectClass=trustedDomain)"):
-            log_info(f'+ {trust["name"]} ({trust["flatName"]})')
-            log_info(
-                f'|___trustAttributes = {list_trust_attributes(trust["trustAttributes"])}'
-            )
-            log_info(
-                f'|___trustDirection = {list_trust_direction(trust["trustDirection"])}'
-            )
-            log_info(f'|___trustType = {list_trust_type(trust["trustType"])}')
-            log_info(f'|___trustPartner = {trust["trustPartner"]}')
-            if "securityIdentifier" in trust:
-                log_info(
-                    f'|___securityIdentifier = {ldap3.protocol.formatters.formatters.format_sid(trust["securityIdentifier"])}'
-                )
-            log_info(f'|___whenCreated = {trust["whenCreated"]}')
-            log_info(f'|___whenChanged = {trust["whenChanged"]}')
-
-    def __print_default_pass_pol(self, pass_pol):
-        """Print info about the default password policy."""
-        log_info("Default password policy:")
-        min_pass_len = pass_pol["minPwdLength"]
-        # Password length
-        if min_pass_len < 8:
-            log_info(f"|___Minimum password length = {c_red(min_pass_len)}")
-        elif min_pass_len < 12:
-            log_info(f"|___Minimum password length = {c_orange(min_pass_len)}")
-        else:
-            log_info(f"|___Minimum password length = {c_green(min_pass_len)}")
-        # Password properties as described here: https://ldapwiki.com/wiki/PwdProperties
-        pass_properties = pass_pol["pwdProperties"]
-        if pass_properties & 1 > 0:
-            log_info(f'|___Password complexity = {c_green("Enabled")}')
-        else:
-            log_info(f'|___Password complexity = {c_red("Disabled")}')
-        # Lockout settings
-        if pass_pol["lockoutThreshold"] == 0:
-            log_success(f'|___Lockout threshold = {c_white_on_red("Disabled")}')
-        else:
-            if pass_pol["lockoutThreshold"] > 5:
-                log_success(f'|___Lockout threshold = {pass_pol["lockoutThreshold"]}')
-            else:
-                log_info(f'|___Lockout threshold = {pass_pol["lockoutThreshold"]}')
-            log_info(
-                f'|___  Lockout duration = {str_human_date(pass_pol["lockoutDuration"])}'
-            )
-            log_info(
-                f'|___  Lockout observation window = {str_human_date(pass_pol["lockOutObservationWindow"])}'
-            )
-        # Password history length
-        if pass_pol["pwdHistoryLength"] > 0:
-            log_success(f'|___Password history length = {pass_pol["pwdHistoryLength"]}')
-        else:
-            log_info(f'|___Password history length = {pass_pol["pwdHistoryLength"]}')
-        # Password min and max age
-        log_info(f'|___Max password age = {str_human_date(pass_pol["maxPwdAge"])}')
-        log_info(f'|___Min password age = {str_human_date(pass_pol["minPwdAge"])}')
-
-    def __print_pass_pol(self, pass_pol):
-        """Print info about a Fine-Grained Password Policy."""
-        log_info(f'Fined grained password policy found: {c_cyan(pass_pol["cn"])}')
-        attributes = pass_pol.entry_attributes_as_dict
-        log_info(
-            f'|____Password settings precedence = {attributes["msDS-PasswordSettingsPrecedence"][0]}'
-        )
-        pass_len = attributes["msDS-MinimumPasswordLength"][0]
-        # Password length
-        if pass_len < 8:
-            log_info(f"|___Minimum password length = {c_red(pass_len)}")
-        elif pass_len < 12:
-            log_info(f"|___Minimum password length = {c_orange(pass_len)}")
-        else:
-            log_info(f"|___Minimum password length = {c_green(pass_len)}")
-        # Password complexity
-        if attributes["msDS-PasswordComplexityEnabled"][0]:
-            log_info(
-                f'|___Password complexity enabled = {c_green(attributes["msDS-PasswordComplexityEnabled"][0])}'
-            )
-        else:
-            log_info(
-                f'|___Password complexity enabled = {c_red(attributes["msDS-PasswordComplexityEnabled"][0])}'
-            )
-        # Password reversible encryption?
-        if attributes["msDS-PasswordReversibleEncryptionEnabled"][0]:
-            log_info(
-                f'|___Password reversible encryption enabled = {c_white_on_red(attributes["msDS-PasswordReversibleEncryptionEnabled"][0])}'
-            )
-        else:
-            log_info(
-                f'|___Password reversible encryption enabled = {attributes["msDS-PasswordReversibleEncryptionEnabled"][0]}'
-            )
-        # Lockout settings
-        if attributes["msDS-LockoutThreshold"][0] == 0:
-            log_success(f'|___Lockout threshold = {c_white_on_red("Disabled")}')
-        else:
-            log_info(
-                f'|___Lockout threshold = {attributes["msDS-LockoutThreshold"][0]}'
-            )
-            log_info(
-                f'|___  Lockout duration = {str_human_date(attributes["msDS-LockoutDuration"][0])}'
-            )
-            log_info(
-                f'|___  Lockout observation window = {str_human_date(attributes["msDS-LockoutObservationWindow"][0])}'
-            )
-
-    def print_pass_pols(self):
-        """Main function to get info about password policies."""
-        # get default password policy
-        default_pps = self.search("(objectClass=domainDNS)", size_limit=5)
-        nb_default_pps = 0
-        for default_pp in default_pps:
-            nb_default_pps += 1
-            self.__print_default_pass_pol(default_pp)
-        if nb_default_pps > 1:
-            log_error(
-                'More than one "default password policy" found. Should not happened. Please investigate and correct the script.'
-            )
-            sys.exit(6)
-        # get Fine Grained Password Policies
-        fgpps = self.search("(objectClass=MsDS-PasswordSettings)", size_limit=100)
-        nb_fgpps = 0
-        for fgpp in fgpps:
-            nb_fgpps += 1
-            self.__print_pass_pol(fgpp)
-        if nb_fgpps <= 0:
-            log_info(
-                "No fine grained password policy found (high privileges are required)."
-            )
-
-    def print_admins(self, size_limit=100):
-        """Method to get a list of members of the "admin" group."""
-        english_groups = "(CN=Administrators)(CN=Domain Admins)(CN=Enterprise Admins)"
-        french_groups = "(CN=Administrateurs)(CN=Admins du domaine)(CN=Administrateurs de l’entreprise)"
-        search_filter = f"(|{english_groups}{french_groups})"
-        # Get the exact distinguishedName of the "admin" group
-        # needed to perform a recursive search of members of members of members ...
-        admin_groups = self.search(
-            search_filter, ["distinguishedName", "cn"], size_limit=10
-        )
-        for admin_group in admin_groups:
-            admins_dn = admin_group["distinguishedName"]
-            log_info(f'All members of group "{admin_group["cn"]}":')
-            # from this distinguishedName, find all members recursively
-            search_filter = f"(&(memberOf:1.2.840.113556.1.4.1941:={admins_dn})(!(objectClass=group)))"
-            attributes = [
-                "objectClass",
-                "name",
-                "userAccountControl",
-                "sAMAccountName",
-                "sAMAccountType",
-            ]
-            admins = self.search(search_filter, attributes, size_limit=size_limit)
-            for admin in admins:
-                self.__print_user_brief(admin, "    ")
-
-    def __print_user_with_spn(self, user):
-        spns = ", ".join(user["servicePrincipalName"])
-        log_success(f"{user['sAMAccountName']} : {spns}")
-
-    def print_kerberoast(self):
-        """Method to get infos about kerberoastable users.
-        Log its sAMAccountName and servicePrincipalName"""
-        search_filter = "(&(objectClass=user)(servicePrincipalName=*)(!(objectClass=computer))(!(cn=krbtgt))(!(userAccountControl:1.2.840.113556.1.4.803:=2)))"
-        search_attributes = ["cn", "samaccountname", "serviceprincipalname"]
-        for kerberoastable_user in self.search(search_filter, search_attributes):
-            self.__print_user_with_spn(kerberoastable_user)
-
-    def print_asreqroast(self):
-        """Method to get all accounts that are vulnerable to ASREPRoast.
-        Filter based on https://www.tarlogic.com/en/blog/how-to-attack-kerberos/"""
-        search_filter = "(&(samAccountType=805306368)(userAccountControl:1.2.840.113556.1.4.803:=4194304))"
-        search_attributes = ["cn", "samaccountname"]
-        for asreqroastuser in self.search(search_filter, search_attributes):
-            log_success(f"{asreqroastuser}")
-
-    def print_search_spn(self, search_filter, size_limit=100):
-        """Method to find services registered in the AD."""
-        if not re.search("serviceprincipalname", search_filter, re.IGNORECASE):
-            search_filter = f"(servicePrincipalName={search_filter}*)"
-        search_attributes = ["cn", "samaccountname", "serviceprincipalname"]
-        for spn_user in self.search(
-            search_filter, search_attributes, size_limit=size_limit
-        ):
-            self.__print_user_with_spn(spn_user)
-
-    def print_lastpwchangekrbtgt(self):
-        """Method to retreive the last time the password for krbtgt was reset."""
-        search_filter = "(cn=krbtgt)"
-        search_attributes = ["whenChanged"]
-        for krbtgt in self.search(search_filter, search_attributes):
-            # when_changed = krbtgt['whenChanged'].replace(microsecond=0).isoformat()
-            # this one is similar but easier to read
-            when_changed = krbtgt["whenChanged"].strftime("%Y-%m-%d %H:%M:%S")
-            log_info(f"krbtgt password changed at {when_changed}")
-
-    def print_search_delegation(self):
-        """Method to retreive accounts with TRUSTED_FOR_DELEGATION set in userAccountControl"""
-        search_filter = "(userAccountControl:1.2.840.113556.1.4.803:=524288)"
-        search_attributes = ["samaccountname"]
-        for account_trusted_for_delegation in self.search(
-            search_filter, search_attributes
-        ):
-            log_success(f"{account_trusted_for_delegation['sAMAccountName']}")
-
-    def print_creator_sid(self):
-        """Main function to get info about createsid from ms-ds-creatorsid."""
-        # get all createsid without parse
-        search_filter = "(ms-ds-creatorsid=*)"
-        search_attributes = ["sAMAccountName", "mS-DS-CreatorSID"]
-        for asreqroastuser in self.search(search_filter, search_attributes):
-            log_success(f"{asreqroastuser}")
-            if asreqroastuser["mS-DS-CreatorSID"] != "":
-                sid_str = convert_sid_to_string(asreqroastuser["mS-DS-CreatorSID"])
-            log_info(sid_str)
-            # parse objectSid
-            search_filter_get_name = f"(objectSid={sid_str})"
-            search_attributes_get_name = ["sAMAccountName"]
-            asreqroastuser_get_name_obj = self.search(
-                search_filter_get_name, search_attributes_get_name
-            )
-            if asreqroastuser_get_name_obj:
-                for asreqroastuser_get_name in asreqroastuser_get_name_obj:
-                    if asreqroastuser_get_name["sAMAccountName"] is not None:
-                        log_info(
-                            f'CreatorSID: {asreqroastuser_get_name["sAMAccountName"]}'
-                        )
-            else:
-                log_error("Maybe already deleted")
-
-    def test_new_search(self):
-        total_entries = 0
-        search_filter = "(samaccountname=it*)"
-        attributes = ["samaccountname"]
-        for entry in self.search(
-            search_filter=search_filter,
-            attributes=attributes,
-            size_limit=22138,
-            page_size=221,
-        ):
-            total_entries += 1
-            # log_info(f"{entry}")
-        log_success(f"We got {total_entries} entries !")
+from ldapsearchad.utils import list_uac_flags
+from ldapsearchad.utils import str_samaccounttype
 
 
 def main():
@@ -895,7 +56,7 @@ def main():
         "-s",
         "--search-filter",
         dest="search_filter",
-        help="Search filter (use LDAP format).",
+        help="Search filter (use LDAP format and don't forget parenthesis).",
     )
     arg_parser.add_argument(
         "search_attributes",
@@ -908,7 +69,7 @@ def main():
         "--size_limit",
         dest="size_limit",
         default=100,
-        help="Size limit (default is 100, or server' own limit).",
+        help="Size limit (default is 100).",
     )
     arg_parser.add_argument(
         "-o",
@@ -932,8 +93,8 @@ def main():
     args = arg_parser.parse_args()
 
     if args.ask_for_version:
-        print_version()
-        sys.exit(0)
+        print(LdapsearchAd.version())
+        exit(0)
 
     # if the version is not asked, we should have at least a target and an action
     if args.ldap_server is None or args.request_type is None:
@@ -956,7 +117,6 @@ def main():
     mandatory_arguments["search-delegation"] = ["domain", "username"]
     mandatory_arguments["createsid"] = ["domain", "username"]
     mandatory_arguments["all"] = ["domain", "username"]
-    mandatory_arguments["test"] = ["domain", "username"]
     actions = [i.strip() for i in args.request_type.split(",")]
     for action in actions:
         if action not in mandatory_arguments.keys():
@@ -971,7 +131,8 @@ def main():
 
     # Configure logging to stdout
     logger = logging.getLogger()
-    handler = logging.StreamHandler(sys.stdout)
+    logger.removeHandler(logger.handlers[0])
+    handler = logging.StreamHandler(stdout)
     if args.verbosity:
         logger.setLevel(logging.DEBUG)
         formatter = logging.Formatter(
@@ -997,20 +158,13 @@ def main():
         args.password,
         args.hashes,
     )
-    if not ldap.server or not ldap.connection:
-        log_error(
-            f"Error, unable to connect to {args.ldap_server} using {args.domain}\\{args.username}"
-        )
-        for error in ldap.last_errors:
-            log_error(error)
-        sys.exit(3)
 
     for action in actions:
 
         # Get the server's infos
         if action == "info":
             log_title("Server infos", 3)
-            ldap.print_info()
+            ldap.infos()
 
         # Check the connection and retrieve the username used for the connection
         elif action == "whoami":
@@ -1105,7 +259,7 @@ def main():
         # Run all checks
         elif action == "all":
             log_title("Server infos", 3)
-            ldap.print_info()
+            ldap.infos()
             log_title('Result of "trusts" command', 3)
             ldap.print_trusts()
             log_title('Result of "pass-pols" command', 3)
@@ -1118,10 +272,6 @@ def main():
             ldap.print_asreqroast()
             log_title('Result of "goldenticket" command', 3)
             ldap.print_lastpwchangekrbtgt()
-
-        elif action == "test":
-            log_title('Result of "test" command', 3)
-            ldap.test_new_search()
 
         else:
             log_error(
