+++ ldapsearchad/__init__.py
@@ -1,0 +1,11 @@
+from .example import add_one
+from .ldapsearchad import LdapsearchAd
+
+
+# version number is just the date of the release.
+# with format: YYYY.MM.DD
+VERSION = "2022.08.17"
+
+
+def version():
+    return VERSION
