+++ ldapsearchad/example.py
@@ -1,0 +1,2 @@
+def add_one(number):
+    return number + 1
