+++ ldapsearchad/exceptions.py
@@ -1,0 +1,5 @@
+class LsaException(Exception):
+    """This is a custom Exception to raise when errors occurs in LdapsearchAd class
+    lsa is the acronym of LdapSearchAd"""
+
+    pass
