+++ setup.py
@@ -1,0 +1,12 @@
+from setuptools import setup
+
+setup(
+    name="ldapsearchad",
+    version="2022.08.18",
+    description="ldapsearch tool to find vulnerable configuration in Active Directory",
+    url="https://github.com/yaap7/ldapsearch-ad",
+    author="yaap7",
+    author_email="g@ywo.fr",
+    license="GPLv3",
+    packages=["ldapsearchad"],
+)
